refactor: house the single-pass x64 code generator under the railshot tier
