+++ bench/suite_test.go
@@ -13,7 +13,7 @@ import (
 	"testing"
 
 	wago "github.com/wago-org/wago"
-	"github.com/wago-org/wago/src/core/compiler/backend/x64"
+	"github.com/wago-org/wago/src/core/compiler/backend/railshot"
 	wasm "github.com/wago-org/wago/src/core/compiler/wasm"
 )
