+++ src/core/compiler/backend/railshot/fold.go
