+++ src/core/compiler/backend/railshot/hints_test.go
